feature: decoupled World from Game
