+++ Rust/002_Grid_World/grid-world/src/main.rs
@@ -1,4 +1,5 @@
 mod game;
+mod world;
 use game::Game;
 use rna::*;
